\name{data}
\alias{data.sim}
\alias{data}
\docType{data}
\title{
Simulated data
}
\description{
This simulated dataset contains three outcomes, and one adjustement variable, divided in two groups.
}
\usage{data(data.sim)}
\format{
  A data frame with 100 observations on the following 5 variables.
  \describe{
    \item{\code{Y1}}{numeric vector of the first outcome}
    \item{\code{Y2}}{numeric vector of the second outcome}
    \item{\code{Y3}}{numeric vector of the third outcome}
    \item{\code{Group}}{binary vector of the group variable}
    \item{\code{A}}{numeric vector of the adjustment variable}
  }
}
\examples{
data(data.sim)
}
\keyword{datasets}
